moved static helper methods to handler interface
